поменял тип данных цены продукта на decimal и пофиксил ошибку, связанную с возвратом общей цены заказа при бронировании

+++ InventoryService/Application/Interfaces/IInventoryService.cs
@@ -10,5 +10,5 @@ public interface IInventoryService
     public Task<Result<List<ProductDto>>> GetProductsAsync();
     public Task<Result<Guid>> CreateProductAsync(CreateProductRequest request);
     public Task<Result<None>> UpdateProductQuantity(UpdateQuantityRequest request, Guid productId);
-    public Task<Result<long>> ReserveItemsAsync(IReadOnlyList<OrderItemDto> itemDtos);
+    public Task<Result<decimal>> ReserveItemsAsync(IReadOnlyList<OrderItemDto> itemDtos);
 }
